+++ src/team.js
@@ -24,7 +24,6 @@ class Team {
         <h3 class="tname">${this.name}</h3>
         <ul class="troster">${this.roster.map(player => `<li id=${player.id}>${player.name}</li>`)}
         </ul>
-        <button class="view-team" data-id=${this.id}>View Team Details</button>
         <button class="delete-team" data-id=${this.id}>Delete Team</button>
         </div>
         `
@@ -44,7 +43,9 @@ class Team {
     }
 
     handleTeamClick(event){
-        teamApiCall.deleteTeam(event)
+        if (event.target.innerText === "Delete Team"){
+            teamApiCall.deleteTeam(event) 
+        }
     }
 
 }
